added elevator interface, added third motor to elbow
