+++ src/main/java/org/sciborgs1155/lib/CustomPeriodRunnables.java
@@ -6,20 +6,20 @@ import java.util.function.BiConsumer;
 
 public final class CustomPeriodRunnables {
 
-    public static final double DEFAULT_RATE = 0.01;
-    private static final Map<Runnable, Double> runnables = new IdentityHashMap<>();
+  public static final double DEFAULT_RATE = 0.01;
+  private static final Map<Runnable, Double> runnables = new IdentityHashMap<>();
 
-    private CustomPeriodRunnables() {}
+  private CustomPeriodRunnables() {}
 
-    public static void add(Runnable runnable, double rate) {
-        runnables.put(runnable, rate);
-    }
+  public static void add(Runnable runnable, double rate) {
+    runnables.put(runnable, rate);
+  }
 
-    public static void add(Runnable runnable) {
-        add(runnable, DEFAULT_RATE);
-    }
+  public static void add(Runnable runnable) {
+    add(runnable, DEFAULT_RATE);
+  }
 
-    public static void forEachRunnable(BiConsumer<? super Runnable, Double> consumer) {
-        runnables.entrySet().forEach(entry -> consumer.accept(entry.getKey(), entry.getValue()));
-    }
+  public static void forEachRunnable(BiConsumer<? super Runnable, Double> consumer) {
+    runnables.entrySet().forEach(entry -> consumer.accept(entry.getKey(), entry.getValue()));
+  }
 }
