+++ src/main/java/org/sciborgs1155/lib/WheelSim.java
@@ -11,44 +11,44 @@ import edu.wpi.first.wpilibj.simulation.DCMotorSim;
 /** Temporary class until wpilib DCMotorSim patch gets upstreamed */
 public class WheelSim extends DCMotorSim {
 
-    public WheelSim(double kV, double kA, DCMotor gearbox, double gearing) {
-        super(identifySystem(kV, kA), gearbox, gearing);
+  public WheelSim(double kV, double kA, DCMotor gearbox, double gearing) {
+    super(identifySystem(kV, kA), gearbox, gearing);
+  }
+
+  /**
+   * Create a state-space model for a 1(?) DOF "wheel" system from its kV (volts/(unit/sec)) and kA
+   * (volts/(unit/sec²). These constants cam be found using SysId. The states of the system are
+   * [position, velocity]ᵀ, inputs are [voltage], and outputs are [position, velocity].
+   *
+   * <p>The distance unit you choose MUST be an SI unit (i.e. meters or radians). You can use the
+   * {@link edu.wpi.first.math.util.Units} class for converting between unit types.
+   *
+   * <p>The parameters provided by the user are from this feedforward model:
+   *
+   * <p>u = K_v v + K_a a
+   *
+   * <p>this is similar to {@link
+   * edu.wpi.first.math.system.plant.LinearSystemId#identifyPositionSystem(double, double)}, but it
+   * uses a (2x2) matrix to preserve position and velocity
+   *
+   * @param kV The velocity gain, in volts/(unit/sec)
+   * @param kA The acceleration gain, in volts/(unit/sec²)
+   * @return A LinearSystem representing the given characterized constants.
+   * @throws IllegalArgumentException if kV &lt;= 0 or kA &lt;= 0.
+   * @see <a href="https://github.com/wpilibsuite/sysid">https://github.com/wpilibsuite/sysid</a>
+   */
+  public static LinearSystem<N2, N1, N2> identifySystem(double kV, double kA) {
+    if (kV <= 0.0) {
+      throw new IllegalArgumentException("Kv must be greater than zero.");
+    }
+    if (kA <= 0.0) {
+      throw new IllegalArgumentException("Ka must be greater than zero.");
     }
 
-    /**
-     * Create a state-space model for a 1(?) DOF "wheel" system from its kV (volts/(unit/sec)) and kA
-     * (volts/(unit/sec²). These constants cam be found using SysId. The states of the system are
-     * [position, velocity]ᵀ, inputs are [voltage], and outputs are [position, velocity].
-     *
-     * <p>The distance unit you choose MUST be an SI unit (i.e. meters or radians). You can use the
-     * {@link edu.wpi.first.math.util.Units} class for converting between unit types.
-     *
-     * <p>The parameters provided by the user are from this feedforward model:
-     *
-     * <p>u = K_v v + K_a a
-     *
-     * <p>this is similar to {@link
-     * edu.wpi.first.math.system.plant.LinearSystemId#identifyPositionSystem(double, double)}, but it
-     * uses a (2x2) matrix to preserve position and velocity
-     *
-     * @param kV The velocity gain, in volts/(unit/sec)
-     * @param kA The acceleration gain, in volts/(unit/sec²)
-     * @return A LinearSystem representing the given characterized constants.
-     * @throws IllegalArgumentException if kV &lt;= 0 or kA &lt;= 0.
-     * @see <a href="https://github.com/wpilibsuite/sysid">https://github.com/wpilibsuite/sysid</a>
-     */
-    public static LinearSystem<N2, N1, N2> identifySystem(double kV, double kA) {
-        if (kV <= 0.0) {
-            throw new IllegalArgumentException("Kv must be greater than zero.");
-        }
-        if (kA <= 0.0) {
-            throw new IllegalArgumentException("Ka must be greater than zero.");
-        }
-
-        return new LinearSystem<N2, N1, N2>(
-                Matrix.mat(Nat.N2(), Nat.N2()).fill(0.0, 1.0, 0.0, -kV / kA),
-                Matrix.mat(Nat.N2(), Nat.N1()).fill(0, 1.0 / kA),
-                Matrix.eye(Nat.N2()),
-                Matrix.mat(Nat.N2(), Nat.N1()).fill(0.0, 0.0));
-    }
+    return new LinearSystem<N2, N1, N2>(
+        Matrix.mat(Nat.N2(), Nat.N2()).fill(0.0, 1.0, 0.0, -kV / kA),
+        Matrix.mat(Nat.N2(), Nat.N1()).fill(0, 1.0 / kA),
+        Matrix.eye(Nat.N2()),
+        Matrix.mat(Nat.N2(), Nat.N1()).fill(0.0, 0.0));
+  }
 }
